Implemented auto.hexe  (ability to flash Hopper programs to MCUs)
Now your Hopper program will persist on your MCU when the power is removed.
Did it properly by implementing File and Directory units using LittleFS.
If auto.hexe is found in /bin, it is launched on restart. Still responds to debugger or monitor connecting and breaking out of current process.