1.0.0.11: Reverting Display drivers
Reworked Wire (I2C) to support multiple I2C controllers in RP2040 MCUs and configurable pins. Also attempting to make a single binary UF2 support as many RP2040 boards as possible (by moving the board-specific code to Hopper source).
Fixed some important defects in optimizer (hence this commit).

Moving display drivers from runtime code to Hopper source (for flexibility and so that we only include the code that is in use in the runtime). Also removed Graphics and Screen syscalls from the Portable Runtime for now.

Work in progress to:
- add decent SPI interface to support TFT LCDs from Hopper code
- restore display drivers
- restore Screen and Graphics support
..